Renamed Client.ts to AdvancedClient.ts.
